+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- fix: Prevent malicious admin from changing JWT Secret Key to masked value. Big thanks to Rapid7 and @M-GRV for the responsible disclosure and @M-GRV for the fix.
 - chore: update Composer deps.
 
 ## [0.4.3] - 2025-06-07